+++ packages/astro/src/vite-plugin-astro/index.ts
@@ -4,10 +4,9 @@ import type { AstroSettings } from '../@types/astro';
 import type { LogOptions } from '../core/logger/core.js';
 import type { PluginMetadata as AstroPluginMetadata } from './types';
 
-import esbuild from 'esbuild';
 import slash from 'slash';
 import { fileURLToPath } from 'url';
-import { cachedCompilation, CompileProps, getCachedSource } from '../core/compile/index.js';
+import { cachedCompilation, CompileProps, getCachedCompileResult } from '../core/compile/index.js';
 import {
 	isRelativePath,
 	prependForwardSlash,
@@ -15,11 +14,13 @@ import {
 	startsWithForwardSlash,
 } from '../core/path.js';
 import { viteID } from '../core/util.js';
-import { getFileInfo, normalizeFilename } from '../vite-plugin-utils/index.js';
+import { normalizeFilename } from '../vite-plugin-utils/index.js';
+import { cachedFullCompilation } from './compile.js';
 import { handleHotUpdate } from './hmr.js';
 import { parseAstroRequest, ParsedRequestResult } from './query.js';
+export { getAstroMetadata } from './metadata.js';
+export type { AstroPluginMetadata };
 
-const FRONTMATTER_PARSE_REGEXP = /^\-\-\-(.*)^\-\-\-/ms;
 interface AstroPluginOptions {
 	settings: AstroSettings;
 	logging: LogOptions;
@@ -103,25 +104,12 @@ export default function astro({ settings, logging }: AstroPluginOptions): vite.P
 			if (!query.astro) {
 				return null;
 			}
-			let filename = parsedId.filename;
-			// For CSS / hoisted scripts we need to load the source ourselves.
-			// It should be in the compilation cache at this point.
-			let raw = await this.resolve(filename, undefined);
-			if (!raw) {
+			// For CSS / hoisted scripts, the main Astro module should already be cached
+			const filename = normalizeFilename(parsedId.filename, config);
+			const compileResult = getCachedCompileResult(config, filename);
+			if (!compileResult) {
 				return null;
 			}
-
-			let source = getCachedSource(config, raw.id);
-			if (!source) {
-				return null;
-			}
-
-			const compileProps: CompileProps = {
-				astroConfig: config,
-				viteConfig: resolvedConfig,
-				filename,
-				source,
-			};
 
 			switch (query.type) {
 				case 'style': {
@@ -129,9 +117,10 @@ export default function astro({ settings, logging }: AstroPluginOptions): vite.P
 						throw new Error(`Requests for Astro CSS must include an index.`);
 					}
 
-					const transformResult = await cachedCompilation(compileProps);
-					const csses = transformResult.css;
-					const code = csses[query.index];
+					const code = compileResult.css[query.index];
+					if (!code) {
+						throw new Error(`No Astro CSS at index ${query.index}`);
+					}
 
 					return {
 						code,
@@ -153,10 +142,7 @@ export default function astro({ settings, logging }: AstroPluginOptions): vite.P
 						};
 					}
 
-					const transformResult = await cachedCompilation(compileProps);
-					const scripts = transformResult.scripts;
-					const hoistedScript = scripts[query.index];
-
+					const hoistedScript = compileResult.scripts[query.index];
 					if (!hoistedScript) {
 						throw new Error(`No hoisted script at index ${query.index}`);
 					}
@@ -171,7 +157,7 @@ export default function astro({ settings, logging }: AstroPluginOptions): vite.P
 						}
 					}
 
-					let result: SourceDescription & { meta: any } = {
+					const result: SourceDescription = {
 						code: '',
 						meta: {
 							vite: {
@@ -182,7 +168,7 @@ export default function astro({ settings, logging }: AstroPluginOptions): vite.P
 
 					switch (hoistedScript.type) {
 						case 'inline': {
-							let { code, map } = hoistedScript;
+							const { code, map } = hoistedScript;
 							result.code = appendSourceMap(code, map);
 							break;
 						}
@@ -215,121 +201,39 @@ export default function astro({ settings, logging }: AstroPluginOptions): vite.P
 				astroConfig: config,
 				viteConfig: resolvedConfig,
 				filename,
+				id,
 				source,
 			};
 
-			try {
-				const transformResult = await cachedCompilation(compileProps);
-				const { fileId: file, fileUrl: url } = getFileInfo(id, config);
+			const transformResult = await cachedFullCompilation({
+				compileProps,
+				rawId: id,
+				logging,
+			});
 
-				for (const dep of transformResult.cssDeps) {
-					this.addWatchFile(dep);
-				}
-
-				// Compile all TypeScript to JavaScript.
-				// Also, catches invalid JS/TS in the compiled output before returning.
-				const { code, map } = await esbuild.transform(transformResult.code, {
-					loader: 'ts',
-					sourcemap: 'external',
-					sourcefile: id,
-					// Pass relevant Vite options, if needed:
-					define: config.vite?.define,
-				});
-
-				let SUFFIX = '';
-				SUFFIX += `\nconst $$file = ${JSON.stringify(file)};\nconst $$url = ${JSON.stringify(
-					url
-				)};export { $$file as file, $$url as url };\n`;
-				// Add HMR handling in dev mode.
-				if (!resolvedConfig.isProduction) {
-					let i = 0;
-					while (i < transformResult.scripts.length) {
-						SUFFIX += `import "${id}?astro&type=script&index=${i}&lang.ts";`;
-						i++;
-					}
-				}
-
-				// Prefer live reload to HMR in `.astro` files
-				if (!resolvedConfig.isProduction) {
-					SUFFIX += `\nif (import.meta.hot) { import.meta.hot.decline() }`;
-				}
-
-				const astroMetadata: AstroPluginMetadata['astro'] = {
-					clientOnlyComponents: transformResult.clientOnlyComponents,
-					hydratedComponents: transformResult.hydratedComponents,
-					scripts: transformResult.scripts,
-				};
-
-				return {
-					code: `${code}${SUFFIX}`,
-					map,
-					meta: {
-						astro: astroMetadata,
-						vite: {
-							// Setting this vite metadata to `ts` causes Vite to resolve .js
-							// extensions to .ts files.
-							lang: 'ts',
-						},
-					},
-				};
-			} catch (err: any) {
-				// Verify frontmatter: a common reason that this plugin fails is that
-				// the user provided invalid JS/TS in the component frontmatter.
-				// If the frontmatter is invalid, the `err` object may be a compiler
-				// panic or some other vague/confusing compiled error message.
-				//
-				// Before throwing, it is better to verify the frontmatter here, and
-				// let esbuild throw a more specific exception if the code is invalid.
-				// If frontmatter is valid or cannot be parsed, then continue.
-				const scannedFrontmatter = FRONTMATTER_PARSE_REGEXP.exec(source);
-				if (scannedFrontmatter) {
-					try {
-						await esbuild.transform(scannedFrontmatter[1], {
-							loader: 'ts',
-							sourcemap: false,
-							sourcefile: id,
-						});
-					} catch (frontmatterErr: any) {
-						// Improve the error by replacing the phrase "unexpected end of file"
-						// with "unexpected end of frontmatter" in the esbuild error message.
-						if (frontmatterErr && frontmatterErr.message) {
-							frontmatterErr.message = frontmatterErr.message.replace(
-								'end of file',
-								'end of frontmatter'
-							);
-						}
-						throw frontmatterErr;
-					}
-				}
-
-				// improve compiler errors
-				if (err.stack && err.stack.includes('wasm-function')) {
-					const search = new URLSearchParams({
-						labels: 'compiler',
-						title: '🐛 BUG: `@astrojs/compiler` panic',
-						template: '---01-bug-report.yml',
-						'bug-description': `\`@astrojs/compiler\` encountered an unrecoverable error when compiling the following file.
-
-**${id.replace(fileURLToPath(config.root), '')}**
-\`\`\`astro
-${source}
-\`\`\``,
-					});
-					err.url = `https://github.com/withastro/astro/issues/new?${search.toString()}`;
-					err.message = `Error: Uh oh, the Astro compiler encountered an unrecoverable error!
-
-    Please open
-    a GitHub issue using the link below:
-    ${err.url}`;
-
-					if (logging.level !== 'debug') {
-						// TODO: remove stack replacement when compiler throws better errors
-						err.stack = `    at ${id}`;
-					}
-				}
-
-				throw err;
+			for (const dep of transformResult.cssDeps) {
+				this.addWatchFile(dep);
 			}
+
+			const astroMetadata: AstroPluginMetadata['astro'] = {
+				clientOnlyComponents: transformResult.clientOnlyComponents,
+				hydratedComponents: transformResult.hydratedComponents,
+				scripts: transformResult.scripts,
+				propagation: 'none',
+			};
+
+			return {
+				code: transformResult.code,
+				map: transformResult.map,
+				meta: {
+					astro: astroMetadata,
+					vite: {
+						// Setting this vite metadata to `ts` causes Vite to resolve .js
+						// extensions to .ts files.
+						lang: 'ts',
+					},
+				},
+			};
 		},
 		async handleHotUpdate(context) {
 			if (context.server.config.isProduction) return;
@@ -337,6 +241,7 @@ ${source}
 				astroConfig: config,
 				viteConfig: resolvedConfig,
 				filename: context.file,
+				id: context.modules[0]?.id ?? undefined,
 				source: await context.read(),
 			};
 			const compile = () => cachedCompilation(compileProps);