+++ packages/astro/test/ssr-request.test.js
@@ -13,6 +13,16 @@ describe('Using Astro.request in SSR', () => {
 			adapter: testAdapter(),
 			output: 'server',
 			base: '/subpath/',
+			integrations: [
+				{
+					name: 'inject-script',
+					hooks: {
+						'astro:config:setup'({ injectScript }) {
+							injectScript('page', 'import "/src/scripts/inject-script.js";');
+						},
+					},
+				},
+			],
 			vite: {
 				build: {
 					assetsInlineLimit: 0,
@@ -64,15 +74,17 @@ describe('Using Astro.request in SSR', () => {
 		const html = await response.text();
 		const $ = cheerioLoad(html);
 
-		const scriptSrc = $('script').attr('src');
-		expect(scriptSrc.startsWith('/subpath/')).to.equal(true);
+		for (const el of $('script')) {
+			const scriptSrc = $(el).attr('src');
+			expect(scriptSrc.startsWith('/subpath/')).to.equal(true);
 
-		request = new Request('http://example.com' + scriptSrc);
-		response = await app.render(request);
+			request = new Request('http://example.com' + scriptSrc);
+			response = await app.render(request);
 
-		expect(response.status).to.equal(200);
-		const js = await response.text();
-		expect(js).to.not.be.an('undefined');
+			expect(response.status).to.equal(200);
+			const js = await response.text();
+			expect(js).to.not.be.an('undefined');
+		}
 	});
 
 	it('assets can be fetched', async () => {